// gccws-ui/src/app/main/module/super-admin/model/sub-report-master.ts
import {CommonObjectProperty} from "../../../core/model/common-object-property";

/**
 * @Author		Emran Khan
 * @Since		  October 01, 2022
 * @version		1.0.0
 */



export class SubReportMaster extends CommonObjectProperty{

  menuItemId?: number;
  menuItemName?: string;
  reportUploadId?: number;
  reportUploadFileName?: string;

}